% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlightR.R
\name{BlightR}
\alias{BlightR}
\title{BlightR model.}
\usage{
BlightR(data, max_na = NULL, temporal_res = "daily",
  model_parameters = "default")
}
\arguments{
\item{data}{The weather data formated as \code{data frame}.}

\item{max_na}{Maximum allowed proportion of missing values. Set to 0.01 by default.}

\item{temporal_res}{By default, the teporal resolution of the output is daily. By changing the argument \code{temporal_res = "hourly"} user will get daily values attached at 12.}

\item{model_parameters}{resoulution of the final data to be returned, daily or hourly. If hourly is selected, outputst are returned at noon.}
}
\value{
This function returns a \code{data.frame} including columns:
\itemize{
 \item \code{"date"}:  Date formated as "yyyy-mm-dd"
 \item \code{"spor"}: Sporulation risk.
 \item \code{"spor_cond"}: Conditions for sporulation.
 \item \code{"inf"}: Infection risk.
 \item \code{"surv_prob"}: Spore mortality.
 \item \code{"risk_si"}: Risk as product of infection and sporulation,
 \item \code{"risk_mi"}: Risk as product of infection and moratlity.
 \item \code{"risk"} Risk as product of sporulation, spore survival and infection.
}
}
\description{
This function calculates potato late blight risk using BlightR model.
The risk needs to be calculated for a single location, so if the
calculation is necessary for multiple locations, data for each location needs to
run separately.
}
\examples{
\donttest{
library(epiCrop)
weather <- system.file("data/blightR_weather.csv", package="epiCrop")
out <- system.time(BlightR(weather))
head(out)
}
}
\keyword{blight,}
\keyword{decision}
\keyword{late}
\keyword{model,}
\keyword{potato}
\keyword{support}
