% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR.R
\name{brown_spot}
\alias{brown_spot}
\title{SEIR}
\usage{
brown_spot(wth, emergence = "2000-05-15", ...)
}
\description{
Adapted from cropsim package version 0.2.0-5 by Adam H.
Sparks - USQ CCH. Original model development: Serge Savary & Rene Pangga.
Original R implementation by Robert J. Hijmans, Rene Pangga, & Jorrel Aunario,
}

