% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_audpc.R
\name{calculate_audpc}
\alias{calculate_audpc}
\title{Calculate the area under the disease progress curve (AUDPC)}
\usage{
calculate_audpc(x)
}
\arguments{
\item{x}{A \link{data.table} object that is the output of \code{\link[=SEIR]{SEIR()}} or any of the
\code{predict} family of functions.}
}
\value{
A \code{numeric} value as \code{double}.
}
\description{
Calculate the area under the disease progress curve (AUDPC)
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# get weather for IRRI Zeigler Experiment Station in wet season 2000
wth <- get_wth(
  lonlat = c(121.25562, 14.6774),
  dates = c("2000-06-30", "2000-12-31")
)
bb <- predict_bacterial_blight(wth, emergence = "2000-07-01")
calculate_audpc(x = bb)
\dontshow{\}) # examplesIf}
}
\references{
Sparks, A.H., P.D. Esker, M. Bates, W. Dall' Acqua, Z. Guo, V. Segovia, S.D.
Silwal, S. Tolos, and K.A. Garrett, 2008. Ecology and Epidemiology in R:
Disease Progress over Time. \emph{The Plant Health Instructor}.
DOI:\link{10.1094/PHI-A-2008-0129-02}https://doi.org/10.1094/PHI-A-2008-0129-02).

Madden, L. V., G. Hughes, and F. van den Bosch. 2007. The Study of Plant
Disease Epidemics. American Phytopathological Society, St. Paul, MN.
DOI:\href{https://doi.org/10.1094/9780890545058}{10.1094/9780890545058}.
}
\author{
Adam H. Sparks, \email{adamhsparks@gmail.com}
}
