% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissolve.R
\name{dissolve}
\alias{dissolve}
\title{Generate summary statistics for epicrop output}
\usage{
dissolve(x)
}
\arguments{
\item{x}{A \link{data.table} object that is the output of \code{\link[=SEIR]{SEIR()}} or any of the
\code{predict} family of functions.  The values provided by this function are the
values used by Savary \emph{et al.} 2012 to report the model outputs in a
\acronym{GIS} framework.}
}
\value{
A \code{\link[data.table:data.table]{data.table::data.table()}} containing the following columns:

\describe{
\item{audpc}{Area under the disease progress curve (\acronym{AUDPC}).}
\item{sd}{Standard deviation of the disease intensity.}
\item{lat}{Latitude value if provided by the \code{x} object.}
\item{lon}{Longitude value if provided by the \code{x} object.}
}
}
\description{
"Dissolves" the voluminous output from \code{\link[=SEIR]{SEIR()}} into easy to interpret
outputs that are useful values for communicating the output of \pkg{epicrop}.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# get weather for IRRI Zeigler Experiment Station in wet season 2000
wth <- get_wth(
  lonlat = c(121.25562, 14.6774),
  dates = c("2000-06-30", "2000-12-31")
)
bb <- predict_bacterial_blight(wth, emergence = "2000-07-01")
dissolve(x = bb)
\dontshow{\}) # examplesIf}
}
\references{
Madden, L. V., G. Hughes, and F. van den Bosch. 2007. The Study of Plant
Disease Epidemics. American Phytopathological Society, St. Paul, MN.
DOI:\href{https://doi.org/10.1094/9780890545058}{10.1094/9780890545058}.

Savary, S., Nelson, A., Willocquet, L., Pangga, I., and Aunario,
J. Modeling and mapping potential epidemics of rice diseases globally. \emph{Crop
Protection}, Volume 34, 2012, Pages 6-17, ISSN 0261-2194 DOI:
\href{https://dx.doi.org/10.1016/j.cropro.2011.11.009}{10.1016/j.cropro.2011.11.009}.

Sparks, A.H., P.D. Esker, M. Bates, W. Dall' Acqua, Z. Guo, V. Segovia, S.D.
Silwal, S. Tolos, and K.A. Garrett, 2008. Ecology and Epidemiology in R:
Disease Progress over Time. \emph{The Plant Health Instructor}.
DOI:\href{https://doi.org/10.1094/PHI-A-2008-0129-02}{10.1094/PHI-A-2008-0129-02}.
}
\author{
Adam H. Sparks, \email{adamhsparks@gmail.com}
}
