% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiCrop.R
\docType{package}
\name{epiCrop}
\alias{epiCrop}
\alias{epiCrop-package}
\title{epiCrop}
\description{
Crop disease epidemiology functions and datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mladencucak/epiCrop}
  \item Report bugs at \url{https://github.com/mladencucak/epiCrop/issues}
}

}
\author{
\strong{Maintainer}: Mladen Cucak \email{mladencucak@gmail.com}

}
