% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epirice-package.R
\docType{package}
\name{epirice-package}
\alias{epirice}
\alias{epirice-package}
\title{epirice: Simulation modeling of rice crop diseases}
\description{
A fork of the 'cropsim' R package by Hijmans et al., this package
    intends to make the EPIRICE model easier to use. The original manuscript by
    Savary et al. (2012) detailing the model use and results modeling global
    epidemics of rice diseases was published in Crop Protection
    (<DOI: 10.1016/j.cropro.2011.11.009>) detailing global disease risk of five
    major rice diseases, bacterial blight, brown spot, leaf blast, sheath blight
    and tungro. `epirice` simulates all of these diseases and also provides a
    generic Susceptible-Exposed-Infectious-Removed model that is suitable for
    use in other pathosystems as well provided the proper inputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adamhsparks/epirice/}
  \item Report bugs at \url{https://github.com/adamhsparks/epirice/issues}
}

}
\author{
\strong{Maintainer}: Adam Sparks \email{adamhsparks@gmail.com} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})

Authors:
\itemize{
  \item Robert Hijmans \email{robert.hijmans@gmail.com} (\href{https://orcid.org/0000-0001-5872-2872}{ORCID}) (Original R Implementation) [translator]
  \item Serge Savary \email{serge.savary@toulouse.inra.fr} (Model idea and development) [conceptor]
  \item Ireno Pangga \email{ibpangga@yahoo.com} (Model Development and Original R Implementation) [translator]
  \item Jorrel Aunario \email{j.aunario@irri.org} (Original R Implementation) [translator]
}

Other contributors:
\itemize{
  \item Aji Sukarta \email{aji.ixb@gmail.com} (Helped fix bug related to SEIR model and establishment
           date.) [contributor]
}

}
\keyword{internal}
