% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatsim.R
\name{getLandCells}
\alias{getLandCells}
\title{Predict brown spot area under the disease progress curve (AUDPC)}
\usage{
getLandCells(tablename, odbcname)
}
\arguments{
\item{wth}{- Weather data source}

\item{emergence}{- Expected date of crop emergence
#' @param \item{...}{Additional arguments - See \link[epiRice]{SEIR} }}
}
\value{
A raster of AUDPC values for brown spot based on given weather data
}
\description{
A dynamic mechanistic simulation of rice brown spot.
The model is driven by daily weather data. Adapted from cropsim package
version 0.2.0-5 by Adam H. Sparks - USQ CCH.
Original model development: Serge Savary & Rene Pangga (IRRI).
Original R implementation by Robert J. Hijmans, Rene Pangga, & Jorrel Aunario
(IRRI).
}
\examples{
wth <- readr::read_table(system.file("extdata", "daily_weather_28368.nasa",
package = "epiRice"), skip = 13)
bs <- brown_spot(wth, onset = 20, duration = 120)
plot(bs, type = 2)

}
\author{
Serge Savary, Ireneo Pangga, Robert Hijmans, Jorrel Khalil Aunario
}
\seealso{
\code{\link[epiRice]{leaf_blast}},
\code{\link[epiRice]{bacterial_blight}}, \code{\link[epiRice]{tungro}}
\code{\link[epiRice]{brown_spot}}, \code{\link[epiRice]{sheath_blight}}
}

