% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlip_rrlex.R
\name{hlip_rrlex}
\alias{hlip_rrlex}
\title{Healthy Latent Infectious Post-Infectious (HLIP) model framework with lesion expansion}
\usage{
hlip_rrlex(
  wth,
  emergence,
  onset,
  duration,
  rhlim,
  rainlim,
  H0,
  I0,
  RcA,
  RcT,
  RcOpt,
  RRLEX,
  p,
  i,
  Sx,
  a,
  RRS,
  RRG
)
}
\arguments{
\item{wth}{a \code{data.frame} of weather on a daily time-step containing data
with the following field names.\tabular{rl}{
   \strong{Field Name} \tab \strong{Value} \cr
   \emph{YYYYMMDD} \tab Date as Year Month Day (ISO8601). \cr
   \emph{DOY} \tab Consecutive day of year, commonly called "Julian date". \cr
   \emph{TEMP} \tab Mean daily temperature (°C). \cr
   \emph{RAIN} \tab Mean daily rainfall (mm). \cr
   \emph{LAT} \tab \strong{Optional} latitude of weather observation. See LAT/LON Note. \cr
   \emph{LON} \tab \strong{Optional} longitude of weather observation. See LAT/LON Note. \cr
}}

\item{emergence}{expected date of plant emergence (or transplanting for rice)
entered in \code{YYYY-MM-DD} format (character).  Described in Table 1 of Savary
\emph{et al.} 2012 and Table 1 of Savary \emph{et al.} 2015.}

\item{onset}{expected number of days until the onset of disease after
emergence date (day, integer).  Described in Table 1 of Savary \emph{et al.}
2012 and Table 1 of Savary \emph{et al.} 2015.}

\item{duration}{simulation duration \emph{i.e.}, growing season length (day,
integer).  Described in Table 1 of Savary \emph{et al.} 2012 and Table 1 of
Savary \emph{et al.} 2015.}

\item{rhlim}{relative humidity value threshold to decide whether leaves are
wet or not (numeric).  Described in Table 1 of Savary \emph{et al.} 2012. Savary
\emph{et al.} 2012 used 90\%.}

\item{rainlim}{rainfall amount (mm) threshold to decide whether leaves are
wet or not (numeric).  Described in Table 1 of Savary \emph{et al.} 2012. Savary
\emph{et al.} 2012 used 5mm.}

\item{H0}{initial number of plant's healthy sites (integer).  Described in
Table 1 of Savary \emph{et al.} 2012 and Table 1 of Savary \emph{et al.} 2015.}

\item{I0}{initial number of infective sites (integer).  Described in Table 1
of Savary \emph{et al.} 2012 and Table 1 of Savary \emph{et al.} 2015.}

\item{RcA}{modifier for \emph{Rc} (the basic infection rate corrected for
removals) for crop age (numeric vector).  Described in Table 1 of Savary
\emph{et al.} 2012 Table 1 of Savary \emph{et al.} 2015.}

\item{RcT}{modifier for \emph{Rc} (the basic infection rate corrected for
removals) for temperature (numeric vector).  Described in Table 1 of Savary
\emph{et al.} 2012 and Table 1 of Savary \emph{et al.} 2015.}

\item{RcOpt}{potential basic infection rate corrected for removals (numeric).
Derived from Table 1 of Savary \emph{et al.} 2012 and Table 1 of Savary \emph{et al.}
2015.}

\item{RRLEX}{relative rate of lesion expansion (numeric).  Described in Table
1 of Savary \emph{et al.} 2015.}

\item{p}{duration of latent period (day, integer).  Described in  Table 1 of
Savary \emph{et al.} 2012 and Table 1 of Savary \emph{et al.} 2015.}

\item{i}{duration of infectious period (day, integer).  Described in Table 1
of Savary \emph{et al.} 2012 and Table 1 of Savary \emph{et al.} 2015.}

\item{Sx}{maximum number of sites (integer).  Described in Table 1 of Savary
\emph{et al.} 2012 and Table 1 of Savary \emph{et al.} 2015.}

\item{a}{aggregation coefficient, values are from 1 to >1 (numeric).
Described in Table 1 of Savary \emph{et al.} 2012 and Table 1 of Savary \emph{et al.}
2015.  See further details in \strong{\emph{a} - Aggregation} section.}

\item{RRS}{relative rate of physiological senescence (numeric).  Described in
Table 1 of Savary \emph{et al.} 2012 and Table 1 of Savary \emph{et al.} 2015.}

\item{RRG}{relative rate of growth (numeric).  Described in Table 1 of Savary
\emph{et al.} 2012 and Table 1 of Savary \emph{et al.} 2015.}
}
\value{
A \code{\link[data.table:data.table]{data.table::data.table()}} containing the following columns:

\describe{
\item{simday}{Zero indexed day of simulation that was run}
\item{dates}{Date of simulation}
\item{sites}{Total number of sites present on day "x"}
\item{latent}{Number of latent sites present on day "x"}
\item{infectious}{Number of infectious sites present on day "x"}
\item{removed}{Number of removed sites present on day "x"}
\item{senesced}{Number of senesced sites present on day "x"}
\item{ratinf}{Rate of infection}
\item{rtransfer}{Rate of transfer from latent to infectious sites}
\item{rgrowth}{Rate of growth of healthy sites}
\item{rsenesced}{Rate of senescence of healthy sites}
\item{rlex}{Rate of lesion expansion}
\item{diseased}{Number of diseased (latent + infectious + removed) sites on
day "x"}
\item{intensity}{Proportion of diseased (latent + infectious + removed)
sites per total sites not including removed sites on day "x"}
\item{lat}{Latitude value if provided by \code{wth} object}
\item{lon}{Longitude value if provided by \code{wth} object}
}
}
\description{
A model framework based on the medical idea of a Susceptible-Exposed-
Infectious-Removed (\acronym{SEIR}) model.  This framework is designed for
plant diseases. This function is originally used by specific disease models
in \sQuote{EPIWHEAT} to model disease intensity of two wheat diseases.  Given
proper values it can be used with other pathosystems as well.
}
\note{
If the \code{wth} object provides \emph{LAT} and \emph{LON} columns, these will be included
in the output for mapping purposes. Both values must be present. These
columns are provided by default when using \code{\link[=get_wth]{get_wth()}}.
}
\section{Differences between \code{hlip()} and \code{hlip_rrlex()}}{

\describe{
\item{\code{RRLEX}}{\code{hlip()}, does not include a parameter for the relative
rate of lesion expansion whereas \code{hlip_rrlex()} does.}
\item{\code{RcW}}{\code{hlip()} uses both relative humidity (RH) and rainfall to
predict leaf wetness.  This differs from \code{hlip_rrlex()}, which uses only
rainfall to predict leaf wetness.}
}
}

\section{\emph{a} - Aggregation}{
When \emph{a} is set to \code{1} the assumption is that that there is no disease
aggregation with new infections occurring at random among the healthy sites.
When \emph{a} is greater than \code{1} there is aggregation in the disease occurrence,
the pathogen is unable to access the entire population of healthy sites,
which results in disease aggregation. Refer to Savary \emph{et al.} (2012) for
greater detail.
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# get weather for Esperance, WA, AU for 2021 growing season
wth <- get_wth(
  lonlat = c(121.891945, -33.861111),
  dates = c("2021-01-01", "2021-12-31")
)

# provide suitable values for brown spot intensity
RcA <-
  cbind(c(0L, 20L, 40L, 60L, 80L, 100L, 120L),
        c(0.35, 0.35, 0.35, 0.47, 0.59, 0.71, 1.0))
RcT <-
  cbind(c(15L, 20L, 25L, 30L, 35L, 40L),
        c(0, 0.06, 1.0, 0.85, 0.16, 0))
emergence <- "2000-07-15"

(hlip_rrlex(
  wth = wth,
  emergence = emergence,
  onset = 20,
  duration = 120,
  rhlim = 90,
  rainlim = 5,
  RcA = RcA,
  RcT = RcT,
  RcOpt = 1.17,
  RRLEX = 0.09,
  p =  11,
  i = 18,
  H0 = 250,
  I0 = 15,
  a = 1,
  Sx = 174000,
  RRS = 0.01,
  RRG = 0.1
))
\dontshow{\}) # examplesIf}
}
\references{
Savary, S., Stetkiewicz, S., Brun, R., and Willocquet, L.
Modelling and Mapping Potential Epidemics of Wheat Diseases—Examples on Leaf
Rust and Septoria Tritici Blotch Using EPIWHEAT. \emph{European Journal of Plant
Pathology} Volume 142, No. 4, 2015, Pages 771–90, DOI:
\url{https://doi.org/10.1007/s10658-015-0650-7}.
}
\seealso{
\code{hlip_rrlex()} is called by the following specific disease modelling functions:
\itemize{
\item \code{\link[=predict_leaf_rust]{predict_leaf_rust()}},
\item \code{\link[=predict_septoria_tritici_blotch]{predict_septoria_tritici_blotch()}}
}
}
\author{
Adam H. Sparks
}
