% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IrishRules.R
\name{IrishRules}
\alias{IrishRules}
\title{Irish Rules}
\usage{
IrishRules(data, model_parameters = "modified", temporal_res = "daily")
}
\arguments{
\item{data}{The weather data in formated as data frame}

\item{model_parameters}{Set of parameters to be used for model run. The default setting
is to use modified version of the model. It is also possible to run the original version of the model.
It is also possible to pass custom parameters for the opurpose of model calibration.}

\item{temporal_res}{By default, the teporal resolution of the output is daily. The output is a data frame
with dates and the estimated risk for the given day. By changing the argument \code{temporal_res = "hourly"}
The output will be returned in an hourly resolution.}
}
\value{
If the desired output resolution is \code{"hourly"} This function returns a \code{data.frame}
including columns:
\itemize{
 \item \code{"date"}:  Date formated as "yyyy-mm-dd"
 \item \code{"ir"}: Daily maximum risk.
}

\itemize{
 \item \code{"date"}:  Date formated as "yyyy-mm-dd"
 \item \code{"hour"}: Hour of the day.
 \item ir Hourly risk.
}
}
\description{
This function calculates potato late blight risk using Modified Irish Ruels model
(Cucak 2019) and the orginal version as reported by Bourke (1953).
}
\examples{
\donttest{
library(epiCrop)
data <- system.file("data/blightR_weather.csv", package="epiCrop")
#Daily output
out <- system.time(IrishRulesModel(data))
head(out)
#Hourly output
out <- system.time(IrishRulesModel(data, temporal_res = "hourly"))
head(out)
}
}
\keyword{Irish}
\keyword{Rules,}
\keyword{blight,}
\keyword{crop}
\keyword{disease}
\keyword{forecasting}
\keyword{late}
\keyword{potato}
