% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR.R
\name{leaf_blast}
\alias{leaf_blast}
\title{Predict leaf blast area under the disease progress curve (AUDPC)}
\usage{
leaf_blast(wth, emergence = "2000-05-15", ...)
}
\arguments{
\item{wth}{- Weather data source}

\item{emergence}{- Expected date of crop emergence}

\item{\item{...}{Additional}{arguments - See \link[epiRice]{SEIR} }}
}
\value{
A raster of AUDPC values for leaf blast based on given weather data
}
\description{
A dynamic mechanistic simulation of leaf blast disease of rice.
The model is driven by daily weather data. Adapted from cropsim package
version 0.2.0-5 by Adam H. Sparks - USQ CCH.
Original model development: Serge Savary & Rene Pangga (IRRI).
Original R implementation by Robert J. Hijmans, Rene Pangga, & Jorrel Aunario
(IRRI).
}
\examples{
wth <- readr::read_table(system.file("extdata", "daily_weather_28368.nasa",
package = "epiRice"), skip = 13)
lb <- leaf_blast(wth, onset = 20, duration = 120)
plot(lb, type = 2)

}
\author{
Serge Savary, Ireneo Pangga, Robert Hijmans, Jorrel Khalil Aunario
}
\references{
Hwang B.K., Koh, Y.J., Chung, H.S., 1987. Effects of adult-plant
resistance on blast severity and yield of rice. Plant Disease 71:1035-1038.

Hemmi, T., Abe, T., Ikaya, J., and Inoue, Y. 1936. Studies on the
rice blast disease. IV. Relation of the environment to the development of
blast disease and physiologic specialization in the rice blast fungus.
Materials for Rural Improvement, Department of Agriculture and Forestry,
Japan No. 105, 145p.

Kato, H and Kozaka, T., 1974. Effect of temperature on lesion
enlargement and sporulation of \emph{Pyricularia oryzae} in rice leaves.
Phytopathology 64:828-830.

Torres, C.Q., 1986. Effect of plant age on the expression of
resistance to \emph{Pyricularia oryzae} Cav. in upland rice varieties. PhD
Thesis, University of the Philippines at Los Banos, 82 p.

El Refaei, M.I., 1977. Epidemiology of rice blast disease in the
tropics with special reference to the leaf wetness in relation to disease
development. PhD Thesis, Indian Agricultural Research Institute, New Delhi,
195 p.

Luo Wei-Hong, 1996. Simulation  and measurement of leaf wetness
formation in paddy rice crops. PhD Thesis, Wageningen Agricultural
University, 87 p.
}
\seealso{
\code{\link[epiRice]{sheath_blight}},
\code{\link[epiRice]{bacterial_blight}}, \code{\link[epiRice]{brown_spot}},
\code{\link[epiRice]{tungro}}
}

