\name{plot}

\docType{methods}

\alias{plot}
\alias{plot,SEIR,missing-method}

\title{ Plotting model results }


\description{
  Function to make a plot of sites, ....
}


\section{Methods}{
\describe{
plot(x, ...)

\tabular{rll}{
\tab \code{x} \tab a RasterLayer object \cr
\tab \code{...} \tab Additional arguments  \cr
}
}
}

\details{
  type =
}

\value{
  None. Only a side effect (the plot)
}

\author{ Robert J. Hijmans }

\keyword{ methods }
