% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_brown_spot.R
\name{predict_brown_spot}
\alias{predict_brown_spot}
\title{Predict rice brown spot severity}
\usage{
predict_brown_spot(wth, emergence, ...)
}
\arguments{
\item{wth}{Weather data with a daily time-step, normally NASA-POWER from
\code{\link{get_wth}}, but any \code{\link[base]{data.frame}} object that has
the following properly named columns in them will work.
\tabular{rl}{
\strong{YYYYMMDD}:\tab Date as Year Month Day (ISO8601).\cr
\strong{DOY}:\tab  Consecutive day of year, commonly called "Julian date".\cr
\strong{TM}:\tab Mean daily temperature (째C).\cr
\strong{TN}:\tab Minimum daily temperature (째C).\cr
\strong{TX}:\tab Maximum daily temperature (째C).\cr
\strong{TDEW}:\tab Mean daily dew point temperature (째C).\cr
\strong{RH}:\tab Mean daily relative humidity (\%).\cr
\strong{RAIN}:\tab Mean daily rainfall (mm).\cr
}}

\item{emergence}{Expected date of crop emergence}

\item{...}{Additional arguments, see \code{\link{SEIR}}}
}
\value{
A \code{\link[data.table]{data.table}} of disease severity and
infection sites. See \code{\link{SEIR}} for a full description of the
column values.
}
\description{
A dynamic mechanistic simulation of rice brown spot, causal agent
\emph{Cochliobolus miyabeanus}. The model is driven by daily weather data,
which can easily be accessed using \code{\link{get_wth}} to download weather
data from \acronym{NASA} \acronym{POWER} using \CRANpkg{nasapower}.
}
\details{
Default values for this disease model are derived from Table 2 (Savary
\emph{et al.} 2012).
}
\note{
Adapted from \pkg{cropsim} package version 0.2.0-5 by Adam H. Sparks - USQ
CCH.
Original model development: Serge Savary & Rene Pangga (IRRI).
Original \R implementation by Robert J. Hijmans, Rene Pangga, & Jorrel
Aunario (IRRI).
}
\examples{
\donttest{
# get weather for IRRI Zeigler Experiment Station in wet season 2000
wth <- get_wth(
  lonlat = c(121.25562, 14.6774),
  dates = c("2000-06-30", "2000-12-31")
)
bs <- predict_brown_spot(wth, emergence = "2000-07-01")
plot(x = bs$dates, y = bs$severity, type = "l")
}

}
\references{
Klomp, A.O., 1977. Early senescence of rice and \emph{Dreschslera
oryzae} in the Wageningen polder, Surinam. PhD thesis, 97p.

Levy, Y. and Cohen, Y., 1980. Sporulation of
\emph{Helminthosporium turcicum} on sweet corn: Effects of temperature and
dew period. Canadian Journal of Plant Pathology 2:65-69.

Padmanabhan, S.Y. and Ganguly, D. 1954. Relation between the age
of rice plant and its susceptibility to \emph{Helminthosporium} and blast
disease. Proceedings of the Indian Academy of Sciences B 29:44-50.

Sarkar, A.K. and Sen Gupta, P.K., 1977. Effect of temperature and
humidity on disease development and sporulation of \emph{Helminthosporium
oryzae} on rice. Indian Phytopathology 30:258-259.



Luo Wei-Hong, 1996. Simulation and measurement of leaf wetness
formation in paddy rice crops. PhD, Wageningen Agricultural University, 87 p.

Waggoner. P.E., Horsfall, J.G., and Lukens, R.J. 1972. EPIMAY. A
Simulator of Southern Corn Leaf Blight. Bulletin of the Connecticut
Experiment Station, New Haven, 85 p.
}
\seealso{
\itemize{
\item \code{\link{SEIR}},
\item \code{\link{predict_bacterial_blight}},
\item \code{\link{predict_leaf_blast}},
\item \code{\link{predict_sheath_blight}}
\item \code{\link{predict_tungro}}
}
}
\author{
Serge Savary, Ireneo Pangga, Robert Hijmans, Jorrel Khalil Aunario
}
