% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_leaf_blast.R
\name{predict_leaf_blast}
\alias{predict_leaf_blast}
\title{Predict rice leaf severity}
\usage{
predict_leaf_blast(wth, emergence, ...)
}
\arguments{
\item{wth}{Weather data with a daily time-step, normally NASA-POWER from
\code{\link{get_wth}}, but any \code{\link[base]{data.frame}} object that has
the following properly named columns in them will work.
\tabular{rl}{
\strong{YYYYMMDD}:\tab Date as Year Month Day (ISO8601).\cr
\strong{DOY}:\tab  Consecutive day of year, commonly called "Julian date".\cr
\strong{TM}:\tab Mean daily temperature (째C).\cr
\strong{TN}:\tab Minimum daily temperature (째C).\cr
\strong{TX}:\tab Maximum daily temperature (째C).\cr
\strong{TDEW}:\tab Mean daily dew point temperature (째C).\cr
\strong{RH}:\tab Mean daily relative humidity (\%).\cr
\strong{RAIN}:\tab Mean daily rainfall (mm).\cr
}}

\item{emergence}{Expected date of crop emergence}

\item{...}{Additional arguments, see \code{\link{SEIR}}}
}
\value{
A \code{\link[data.table]{data.table}} of disease severity and
infection sites. See \code{\link{SEIR}} for a full description of the
column values.
}
\description{
A dynamic mechanistic simulation of leaf blast disease of rice, causal agent
\emph{Magnaporthe oryzae}.
The model is driven by daily weather data. Adapted from \pkg{cropsim} package
version 0.2.0-5 by Adam H. Sparks - USQ CCH.
Original model development: Serge Savary & Rene Pangga (IRRI).
Original \R implementation by Robert J. Hijmans, Rene Pangga, & Jorrel
Aunario (IRRI).
}
\examples{
\donttest{
# get weather for IRRI Zeigler Experiment Station in dry season 2000
wth <- get_wth(
  lonlat = c(121.25562, 14.6774),
  dates = c("2000-05-15", "2000-12-31")
)
lb <- predict_leaf_blast(wth, emergence = "2000-05-18")
plot(x = lb$dates, y = lb$severity, type = "l")
}
}
\references{
Hwang B.K., Koh, Y.J., Chung, H.S., 1987. Effects of adult-plant
resistance on blast severity and yield of rice. Plant Disease 71:1035-1038.

Hemmi, T., Abe, T., Ikaya, J., and Inoue, Y. 1936. Studies on the
rice blast disease. IV. Relation of the environment to the development of
blast disease and physiologic specialization in the rice blast fungus.
Materials for Rural Improvement, Department of Agriculture and Forestry,
Japan No. 105, 145p.

Kato, H and Kozaka, T., 1974. Effect of temperature on lesion
enlargement and sporulation of \emph{Pyricularia oryzae} in rice leaves.
Phytopathology 64:828-830.

Torres, C.Q., 1986. Effect of plant age on the expression of
resistance to \emph{Pyricularia oryzae} Cav. in upland rice varieties. PhD
Thesis, University of the Philippines at Los Banos, 82 p.

El Refaei, M.I., 1977. Epidemiology of rice blast disease in the
tropics with special reference to the leaf wetness in relation to disease
development. PhD Thesis, Indian Agricultural Research Institute, New Delhi,
195 p.

Luo Wei-Hong, 1996. Simulation  and measurement of leaf wetness
formation in paddy rice crops. PhD Thesis, Wageningen Agricultural
University, 87 p.
}
\seealso{
\itemize{
\item \code{\link{SEIR}},
\item \code{\link{predict_bacterial_blight}},
\item \code{\link{predict_brown_spot}},
\item \code{\link{predict_sheath_blight}}
\item \code{\link{predict_tungro}}
}
}
\author{
Serge Savary, Ireneo Pangga, Robert Hijmans, Jorrel Khalil Aunario
}
