% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_leaf_rust.R
\name{predict_leaf_rust}
\alias{predict_leaf_rust}
\title{Predict wheat leaf rust severity}
\usage{
predict_leaf_rust(wth, emergence, ...)
}
\arguments{
\item{wth}{Weather data with a daily time-step, normally NASA POWER from
\code{\link{get_wth}}, but any \code{\link[base]{data.frame}} object that has
the following properly named columns in them will work.
\tabular{rl}{
\strong{YYYYMMDD}:\tab Date as Year Month Day (ISO8601).\cr
\strong{DOY}:\tab  Consecutive day of year, commonly called "Julian date".\cr
\strong{TEMP}:\tab Mean daily temperature (°C).\cr
\strong{TMIN}:\tab Minimum daily temperature (°C).\cr
\strong{TMAX}:\tab Maximum daily temperature (°C).\cr
\strong{TDEW}:\tab Mean daily dew point temperature (°C).\cr
\strong{RHUM}:\tab Mean daily temperature (°C).\cr
\strong{RAIN}:\tab Mean daily rainfall (mm).\cr
}}

\item{emergence}{Expected date of crop emergence}

\item{...}{Additional arguments, see \code{\link{SEIR}}}
}
\value{
A \code{\link[data.table]{data.table}} of disease severity and
infection sites. See \code{\link{SEIR}} for a full description of the
column values.
}
\description{
A dynamic mechanistic simulation of wheat leaf rust, causal agent,
\emph{Puccinia tritici}. The model is driven by daily weather data,
which can easily be accessed using \code{\link{get_wth}} to download weather
data from \acronym{NASA} \acronym{POWER} using \CRANpkg{nasapower}.
}
\details{
Default values for this disease model are derived from Table 2 (Savary
\emph{et al.} 2015).
}
\examples{
\donttest{
# get weather for Manhattan, Kansas
wth <- get_wth(
  lonlat = c(-96.5717, 39.1836),
  dates = c("2000-08-01", "2001-06-30")
)
bs <- predict_leaf_rust(wth, emergence = "2000-08-15")
plot(x = bs$dates, y = bs$severity, type = "l")
}

}
\references{
Savary, S., Nelson, A., Willocquet, L., Pangga, I., and Aunario, J. Modeling
and mapping potential epidemics of rice diseases globally. Crop Protection,
Volume 34, 2012, Pages 6-17, ISSN 0261-2194 DOI:
\url{https://doi.org/10.1016/j.cropro.2011.11.009}.

Savary, S., Stetkiewicz, S., Brun, F., and Willocquet, L. Modelling and
Mapping Potential Epidemics of Wheat Diseases—Examples on Leaf Rust and
Septoria Tritici Blotch Using EPIWHEAT. European Journal of Plant Pathology
142, no. 4 (August 1, 2015): 771–90. DOI:
\url{https://doi.org/10.1007/s10658-015-0650-7}.
}
\seealso{
\itemize{
\item \code{\link{SEIR}},
\item \code{\link{predict_bacterial_blight}},
\item \code{\link{predict_brown_spot}},
\item \code{\link{predict_leaf_blast}},
\item \code{\link{predict_sheath_blight}}
\item \code{\link{predict_tungro}}
}
}
\author{
Adam H. Sparks, \email{adamhsparks@gmail.com}
}
