% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_tungro.R
\name{predict_tungro}
\alias{predict_tungro}
\title{Predict rice tungro severity}
\usage{
predict_tungro(wth, emergence, ...)
}
\arguments{
\item{wth}{Weather data with a daily time-step, normally NASA-POWER from
\code{\link{get_wth}}, but any \code{\link[base]{data.frame}} object that has
the following properly named columns in them will work.
\tabular{rl}{
\strong{YYYYMMDD}:\tab Date as Year Month Day (ISO8601).\cr
\strong{DOY}:\tab  Consecutive day of year, commonly called "Julian date".\cr
\strong{TM}:\tab Mean daily temperature (째C).\cr
\strong{TN}:\tab Minimum daily temperature (째C).\cr
\strong{TX}:\tab Maximum daily temperature (째C).\cr
\strong{TDEW}:\tab Mean daily dew point temperature (째C).\cr
\strong{RH}:\tab Mean daily relative humidity (\%).\cr
\strong{RAIN}:\tab Mean daily rainfall (mm).\cr
}}

\item{emergence}{Expected date of crop emergence}

\item{...}{Additional arguments, see \code{\link{SEIR}}}
}
\value{
A \code{\link[data.table]{data.table}} of disease severity and
infection sites. See \code{\link{SEIR}} for a full description of the
column values.
}
\description{
A dynamic mechanistic simulation of tungro disease of rice, causal agents
\emph{Rice Tungro Spherical Virus} and \emph{Rice Tungro Baciliform Virus}.
The model is driven by daily weather data. Adapted from \pkg{cropsim} package
version 0.2.0-5 by Adam H. Sparks - USQ CCH.
Original model development: Serge Savary & Rene Pangga (IRRI).
Original \R implementation by Robert J. Hijmans, Rene Pangga, & Jorrel
Aunario (IRRI).
}
\examples{
\donttest{
# get weather for IRRI Zeigler Experiment Station in wet season 2000
wth <- get_wth(
  lonlat = c(121.25562, 14.6774),
  dates = c("2000-06-30", "2000-12-31")
)
tg <- predict_tungro(wth, emergence = "2000-07-01")
plot(x = tg$dates, y = tg$severity, type = "l")
}
}
\references{
Tiongco, E.R., Cabunagan, R.C., Flores, Z.M., Hibino, H., and
Koganezawa, H., 1993. Serological monitoring of rice tungro disease
development in the field: its implication in disease management.
Plant Disease 77:877-882.

Rivera, C.T. and Ou, S.H., 1965. Leafhopper transmission of
tungro disease of rice. Plant Disease Reporter 49:127-131.

Ling, K.C., Palomar, M.K., 1966. Studies on rice plants infected
with the tungro virus at different ages. Philippines Agriculturist 50:165-177.

Ling, K.C., and Tiongco, E.R., 1976. Effect of temperature on the
transmission of rice tungro virus by \emph{Nephotettix virescens}.
Philippine Phytopathology 11:46-57.
}
\seealso{
\itemize{
\item \code{\link{SEIR}},
\item \code{\link{predict_bacterial_blight}},
\item \code{\link{predict_brown_spot}},
\item \code{\link{predict_leaf_blast}}
\item \code{\link{predict_sheath_blight}}
}
}
\author{
Serge Savary, Ireneo Pangga, Robert Hijmans, Jorrel Khalil Aunario
}
