% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR.R
\name{sheath_blight}
\alias{sheath_blight}
\title{Predict sheath blight area under the disease progress curve (AUDPC)}
\usage{
sheath_blight(wth, emergence = "2000-05-15", ...)
}
\arguments{
\item{wth}{- Weather data source}

\item{emergence}{- Expected date of crop emergence}

\item{...}{Additional arguments - See \link[epirice]{SEIR}}
}
\value{
A raster of AUDPC values for sheath blight based on given weather
data
}
\description{
A dynamic mechanistic simulation of sheath blight disease of rice.
The model is driven by daily weather data. Adapted from cropsim package
version 0.2.0-5 by Adam H. Sparks - USQ CCH.
Original model development: Serge Savary & Rene Pangga (IRRI).
Original R implementation by Robert J. Hijmans, Rene Pangga, & Jorrel Aunario
(IRRI).
}
\examples{
wth <- readr::read_table(system.file("extdata", "daily_weather_28368.nasa",
package = "epirice"), skip = 13)
sb <- sheath_blight(wth, onset = 20, duration = 120)
plot(bb, type = 2)

}
\references{
Savary, S., Castilla, N.P., Willocquet, L. 2001. Analysis of the
spatio-temporal structure of rice sheath blight epidemics in a farmer's
field. Plant Pathology 50:53-68.

Savary, S., Willocquet, L., Teng, P.S., 1997. Modelling sheath
blight epidemics on rice tillers. Agricultural Systems 55:359-384.

Castilla, N.P., Leano, R.M., Elazegui, F.A., Teng, P.S., Savary,
S., 1996. Effects of plant contacts, inoculation pattern, leaf wetness
regime, and nitrogen supply on inoculum efficiency in rice sheath blight.
Journal of Phytopathology 144:187-192.

Sharma, N.R., Teng, P.S., Olivares, F.M., 1990. Effect of rice
growth stage on sheath blight (ShB) development and yield loss. International
Rice Research Newsletter 15:19-20.

Hashiba, T. and Ijiri, T., 1989. Estimation of yield loss and
computerized forecasting system (BLIGHTAS) for rice sheath blight disease.
International Symposium on Tropical Agricultural Research: Crop losses due to
disease outbreaks in the tropics and countermeasures. Tropical Agricultural
Research Series (Japan) No. 22 pp. 163-171.

Tu, C.C., Chang, Y.C., Wang, C.W., 1979. Studies on the ecology
of \emph{Rhizoctonia solani}, the causal organism of rice sheath blight.
National Science Council Monthly, ROC 7:1208-1219.

Gross, M.K., Santini, J.B., Tikhonova, I. and Latin, R. 1998. The
influence of temperature and leaf wetness duration on infection of perennial
ryegrass by \emph{Rhizoctonia solani}. Plant Disease 82:1012-1016.
}
\seealso{
\code{\link[epirice]{leaf_blast}},
\code{\link[epirice]{bacterial_blight}}, \code{\link[epirice]{brown_spot}},
\code{\link[epirice]{tungro}}
}
\author{
Serge Savary, Ireneo Pangga, Robert Hijmans, Jorrel Khalil Aunario
}
