% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR.R
\name{tungro}
\alias{tungro}
\title{Predict tungro area under the disease progress curve (AUDPC)}
\usage{
tungro(wth, emergence = "2000-05-15", ...)
}
\arguments{
\item{wth}{- Weather data source}

\item{emergence}{- Expected date of crop emergence}

\item{...}{Additional arguments - See \link[epirice]{SEIR}}
}
\value{
A raster of AUDPC values for tungro based on given weather data
}
\description{
A dynamic mechanistic simulation of tungro disease of rice.
The model is driven by daily weather data. Adapted from cropsim package
version 0.2.0-5 by Adam H. Sparks - USQ CCH.
Original model development: Serge Savary & Rene Pangga (IRRI).
Original R implementation by Robert J. Hijmans, Rene Pangga, & Jorrel Aunario
(IRRI).
}
\examples{
wth <- readr::read_table(system.file("extdata", "daily_weather_28368.nasa",
package = "epirice"), skip = 13)
tg <- tungro(wth, onset = 20, duration = 120)
plot(tg, type = 2)

}
\references{
Tiongco, E.R., Cabunagan, R.C., Flores, Z.M., Hibino, H., and
Koganezawa, H., 1993. Serological monitoring of rice tungro disease
development in the field: its implication in disease management.
Plant Disease 77:877-882.

Rivera, C.T. and Ou, S.H., 1965. Leafhopper transmission of
tungro disease of rice. Plant Disease Reporter 49:127-131.

Ling, K.C., Palomar, M.K., 1966. Studies on rice plants infected
with the tungro virus at different ages. Philippines Agriculturist 50:165-177.

Ling, K.C., and Tiongco, E.R., Effect of temperature on the
transmission of rice tungro virus by \emph{Nephotettix virescens}.
Philippine Phytopathology 11:46-57.
}
\seealso{
\code{\link[epirice]{leaf_blast}},
\code{\link[epirice]{bacterial_blight}}, \code{\link[epirice]{brown_spot}},
\code{\link[epirice]{sheath_blight}}
}
\author{
Serge Savary, Ireneo Pangga, Robert Hijmans, Jorrel Khalil Aunario
}
