% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WthCheckImpute.R
\name{WthCheckImpute}
\alias{WthCheckImpute}
\title{Weather interpolation}
\usage{
WthCheckImpute(data, report = TRUE)
}
\arguments{
\item{data}{The weather data formated as \code{data frame}.}

\item{report}{Maximum allowed proportion of missing values. Set to 0.01 by default.}
}
\value{
This function returns a \code{data.frame} including columns:
}
\description{
This function calculates potato late blight risk using BlightR model.
}
\examples{
\donttest{
library(epiCrop)
weather[666:674,c("temp","rhum")] <- NA #9 NAs
weather[766:775,c("temp","rhum")] <- NA #10 NAs
weather[866:876,c("temp","rhum")] <- NA #11 NAs
weather[, "sol_rad"] <- NA
weather[5,c("temp","rhum", "rain")] <- 1000 #large values
weather[10,c("temp","rhum", "rain")] <- -9999 #negative
WthCheckImpute(weather)
}
}
\keyword{interpolation}
\keyword{weather}
