% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiCrop.R
\docType{package}
\name{epiCrop}
\alias{epiCrop}
\alias{epiCrop-package}
\title{epiCrop}
\description{
Crop disease epidemiology functions and datasets.
}
\author{
\strong{Maintainer}: Mladen Cucak \email{mladencucak@gmail.com}

}
