% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epirice-package.R
\docType{package}
\name{epirice-package}
\alias{epirice}
\alias{epirice-package}
\title{epirice: Simulation modeling of rice crop diseases}
\description{
A fork of the 'cropsim' R package by Hijmans et al. This package is
    intended to make the EPIRICE model easier to use while stripping other
    functions not related to rice diseases from the R package and removing
    the hooks to the IRRI GIS lab databases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adamhsparks/epirice}
}

}
\author{
\strong{Maintainer}: Adam Sparks \email{adamhsparks@gmail.com} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})

Authors:
\itemize{
  \item Robert Hijmans \email{robert.hijmans@gmail.com} (Original R Implementation) [translator]
  \item Serge Savary \email{serge.savary@toulouse.inra.fr} (Model idea and development) [conceptor]
  \item Ireno Pangga \email{ibpangga@yahoo.com} (Model Development and Original R Implementation) [translator]
  \item Jorrel Aunario \email{j.aunario@irri.org} (Original R Implementation) [translator]
}

Other contributors:
\itemize{
  \item Aji Sukarta \email{aji.ixb@gmail.com} (Helped fix bug related to SEIR model and establishment
           date.) [contributor]
}

}
\keyword{internal}
