% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epirice-package.R
\docType{package}
\name{epirice-package}
\alias{epirice}
\alias{epirice-package}
\title{epirice: Simulation Modelling of Rice Crop Diseases Using a Susceptible-Exposed-Infectious-Removed (SEIR) Model}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A fork of the 'cropsim' R package by Hijmans et al., this package
    intends to make the 'EPIRICE' model easier to use. The original manuscript
    by Savary et al. (2012) <doi:10.1016/j.cropro.2011.11.009> detailed the
    model's development, use and results from modeling global unmanaged
    epidemics of rice diseases for five major rice diseases, bacterial blight,
    brown spot, leaf blast, sheath blight and tungro. `epirice` provides
    functions to simulate all of these diseases and also provides a generic
    Susceptible-Exposed-Infectious-Removed ('SEIR') model that is suitable for
    use in other pathosystems as well when provided with the proper inputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adamhsparks/epirice/}
  \item Report bugs at \url{https://github.com/adamhsparks/epirice/issues}
}

}
\author{
\strong{Maintainer}: Adam Sparks \email{adamhsparks@gmail.com} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})

Authors:
\itemize{
  \item Robert Hijmans \email{robert.hijmans@gmail.com} (\href{https://orcid.org/0000-0001-5872-2872}{ORCID}) (Original R implementation) [translator]
  \item Serge Savary \email{serge.savary@toulouse.inra.fr} (Model conception and development in Stella) [conceptor]
  \item Ireno Pangga \email{ibpangga@yahoo.com} (Model development and original R implementation) [translator]
  \item Jorrel Aunario \email{j.aunario@irri.org} (Original R implementation) [translator]
}

Other contributors:
\itemize{
  \item Aji Sukarta \email{aji.ixb@gmail.com} (Helped fix bug related to SEIR model and establishment
           date.) [contributor]
  \item Emerson Del Ponte \email{delponte@ufv.br} (\href{https://orcid.org/0000-0003-4398-409X}{ORCID}) (Suggested fixes for several typos in the documentation) [contributor]
}

}
\keyword{internal}
