% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wth.R
\name{get_wth}
\alias{get_wth}
\title{Get Weather Data from NASA POWER}
\usage{
get_wth(lonlat, dates)
}
\arguments{
\item{lonlat}{A numeric vector of geographic coordinates for a cell or region
entered as x, y coordinates.}

\item{dates}{A character vector of start and end dates in that order.}
}
\value{
A \code{data.frame} of weather data suitable for use in \pkg{epirice}.
}
\description{
Get Weather Data from NASA POWER
}
\examples{
wth <- get_wth(lonlat = c(151.81, -27.48),
               dates = c("2015-01-15", "2015-05-15"))
}
