% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wth.R
\name{get_wth}
\alias{get_wth}
\title{Get weather data from NASA POWER}
\usage{
get_wth(lonlat, dates)
}
\arguments{
\item{lonlat}{A numeric vector of geographic coordinates for a cell or region
entered as x, y coordinates.}

\item{dates}{A character vector of start and end dates in that order.}
}
\value{
A \code{\link[data.table]{data.table}} of weather data suitable for
use in \pkg{epirice} with the following columns:
\tabular{rl}{
\strong{YYYYMMDD}:\tab Date as Year Month Day (ISO8601).\cr
\strong{DOY}:\tab  Consecutive day of year, commonly called "Julian date".\cr
\strong{TM}:\tab Mean daily temperature (째C).\cr
\strong{TN}:\tab Minimum daily temperature (째C).\cr
\strong{TX}:\tab Maximum daily temperature (째C).\cr
\strong{TDEW}:\tab Mean daily dew point temperature (째C).\cr
\strong{RH}:\tab Mean daily relative humidity (\%).\cr
\strong{RAIN}:\tab Mean daily rainfall (mm).\cr
}
}
\description{
Get weather data from NASA POWER
}
\details{
This function is just a wrapper for the \CRANpkg{nasapower}
\code{\link[nasapower]{get_power}} function with predefined parameters
suitable for use in EPIRICE.
}
\examples{
\donttest{
# get weather for IRRI Zeigler Experiment Station in dry season 2000
wth <- get_wth(
  lonlat = c(121.25562, 14.6774),
  dates = c("2000-01-15", "2000-05-31")
)
}
}
\author{
Adam H. Sparks
}
