% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacterial_blight.R
\name{predict_bacterial_blight}
\alias{predict_bacterial_blight}
\title{Predict rice bacterial blight area under the disease progress curve (AUDPC)}
\usage{
predict_bacterial_blight(wth, emergence, ...)
}
\arguments{
\item{wth}{Weather data with a daily time-step, normally NASA-POWER from
\code{\link{get_wth}}, but any data that has the properly named fields data
\itemize{
\item YYYYMMDD Date in YYYY-MM-DD format
\item DOY Numeric day of year, e.g. 1 - 365
\item T2M Mean daily temperature
\item T2MN Minimum daily temperature
\item T2MX Maximum daily temperature
\item RH2M Relative humidity
\item RAIN Precipitation
}}

\item{emergence}{Expected date of crop emergence}

\item{...}{Additional arguments, see \code{\link{SEIR}}}
}
\value{
An \pkg{epirice} \code{SEIR} object
}
\description{
A dynamic mechanistic simulation of bacterial blight disease of rice,
causal agent \emph{Xanthomonas oryzae} pv. \emph{oryzae}.
The model is driven by daily weather data. Adapted from \pkg{cropsim} package
version 0.2.0-5 by Adam H. Sparks - USQ CCH.
Original model development: Serge Savary & Rene Pangga (IRRI).
Original \R implementation by Robert J. Hijmans, Rene Pangga, & Jorrel
Aunario (IRRI).
}
\examples{
\donttest{
# get weather for IRRI Zeigler Experiment Station in dry season 2000
wth <- get_wth(
  lonlat = c(121.25562, 14.6774),
  dates = c("2000-05-15", "2000-12-31")
)
bb <- predict_bacterial_blight(wth, emergence = "2000-05-18")
}

}
\references{
Adhikari, T.B., 1991. Effects of rice genotype and environment on
bacterial blight progression. PhD thesis, University of the Philippines at
Los Baños, 143 p.

Nayak, P., Suriya Rao, A.V., Chakrabarti, N.K., 1987. Components
of resistance to bacterial blight disease of rice. Journal of Phytopathology
119:312-318.

Baw A. and Mew, T.W., 1988. Scoring systems for evaluating rice
varietal resistance to bacterial blight (BB): lesion size by growth stage.
International Rice Research Newsletter 13:10-11.

Luo Wei-Hong, 1996. Simulation and measurement of leaf wetness
formation in paddy rice crops. PhD, Wageningen Agricultural University, 87 p.

Horino, O., Mew, T.W., Yamada, T., 1982. The effect of
temperature on the development of bacterial leaf blight on rice. Annals of
the Phytopathological Society of Japan 48: 72-75

Medalla, E. 1992. Characterization of resistance of IR cultivars
to two races of \emph{Xanthomonas oryzae} pv. \emph{oryzae}. Unpublished M.S.
Thesis, University of the Philippines at Los Baños, 81 p.
}
\seealso{
\code{\link{predict_leaf_blast}},
\code{\link{predict_bacterial_blight}}, \code{\link{predict_brown_spot}},
\code{\link{predict_sheath_blight}}
}
\author{
Serge Savary, Ireneo Pangga, Robert Hijmans, Jorrel Khalil Aunario
}
