% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_brown_spot.R
\name{predict_brown_spot}
\alias{predict_brown_spot}
\title{Predict rice brown spot severity}
\usage{
predict_brown_spot(wth, emergence, ...)
}
\arguments{
\item{wth}{Weather data with a daily time-step, normally NASA POWER from
\code{\link[=get_wth]{get_wth()}}, but any\link[=data.frame]{base} object that has
the following properly named columns in them will work.\tabular{rl}{
   \strong{Field Name} \tab \strong{Value} \cr
   \emph{YYYYMMDD} \tab Date as Year Month Day (ISO8601) \cr
   \emph{DOY} \tab Consecutive day of year, commonly called "Julian date" \cr
   \emph{TEMP} \tab Mean daily temperature (°C) \cr
   \emph{RHUM} \tab Mean daily temperature (°C) \cr
   \emph{RAIN} \tab Mean daily rainfall (mm) \cr
}}

\item{emergence}{Expected date of crop emergence}

\item{...}{Additional arguments, see \code{\link[=SEIR]{SEIR()}}}
}
\value{
A \link[=data.table]{data.table} of disease severity and
infection sites. See \code{\link[=SEIR]{SEIR()}} for a full description of the
column values.
}
\description{
A dynamic mechanistic simulation of rice brown spot, causal agent
\emph{Cochliobolus miyabeanus}. The model is driven by daily weather data, which
can easily be accessed using\code{\link[=get_wth]{get_wth()}} to download weather data from
\acronym{NASA} \acronym{POWER} using \CRANpkg{nasapower}.
}
\details{
Default values for this disease model are derived from Table 2 (Savary
\emph{et al.} 2012).
}
\note{
Adapted from \pkg{cropsim} package version 0.2.0-5 by Adam H. Sparks,
University of Southern Queensland Centre for Crop Health.
Original model development: Serge Savary & Rene Pangga (IRRI).
Original \R implementation by Robert J. Hijmans, Rene Pangga, & Jorrel
Aunario (IRRI).

The optimum temperature for brown spot as presented in Table 2 of
Savary \emph{et al.} 2012 has a typo. The optimal value should be 25 °C, not
20 °C as shown. The correct value, 25 °C, is used in this implementation.
}
\examples{
\donttest{
# get weather for IRRI Zeigler Experiment Station in wet season 2000
wth <- get_wth(
  lonlat = c(121.25562, 14.6774),
  dates = c("2000-06-30", "2000-12-31")
)
bs <- predict_brown_spot(wth, emergence = "2000-07-01")
plot(x = bs$dates, y = bs$severity, type = "l")
}

}
\references{
Klomp, A.O., 1977. Early senescence of rice and \emph{Drechslera
oryzae} in the Wageningen polder, Surinam. PhD thesis, 97p.

Levy, Y. and Cohen, Y., 1980. Sporulation of \emph{Helminthosporium turcicum} on
sweet corn: Effects of temperature and dew period. Canadian Journal of Plant
Pathology 2:65-69. DOI: \url{https://doi.org/10.1080/07060668009501440}.

Luo Wei-Hong, 1996. Simulation and measurement of leaf wetness formation in
paddy rice crops. PhD, Wageningen Agricultural University, 87 p.

Padmanabhan, S.Y. and Ganguly, D. 1954. Relation between the age of rice
plant and its susceptibility to \emph{Helminthosporium} and blast disease.
Proceedings of the Indian Academy of Sciences B 29:44-50.

Sarkar, A.K. and Sen Gupta, P.K., 1977. Effect of temperature and humidity on
disease development and sporulation of \emph{Helminthosporium oryzae} on rice.
Indian Phytopathology 30:258-259.

Savary, S., Nelson, A., Willocquet, L., Pangga, I., and Aunario, J. Modeling
and mapping potential epidemics of rice diseases globally. Crop Protection,
Volume 34, 2012, Pages 6-17, ISSN 0261-2194 DOI:
\url{https://doi.org/10.1016/j.cropro.2011.11.009}.

Waggoner. P.E., Horsfall, J.G., and Lukens, R.J. 1972. EPIMAY. A Simulator of
Southern Corn Leaf Blight. Bulletin of the Connecticut Experiment Station,
New Haven, 85 p.
}
\seealso{
\itemize{
\item \code{\link[=SEIR]{SEIR()}},
\item \code{\link[=predict_bacterial_blight]{predict_bacterial_blight()}},
\item \code{\link[=predict_leaf_blast]{predict_leaf_blast()}},
\item \code{\link[=predict_sheath_blight]{predict_sheath_blight()}},
\item \code{\link[=predict_tungro]{predict_tungro()}}
}
}
\author{
Serge Savary, Ireneo Pangga, Robert Hijmans, Jorrel Khalil Aunario
}
