% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf_blast.R
\name{predict_leaf_blast}
\alias{predict_leaf_blast}
\title{Predict rice leaf blast area under the disease progress curve (AUDPC)}
\usage{
predict_leaf_blast(wth, emergence, ...)
}
\arguments{
\item{wth}{Weather data with a daily time-step, normally NASA-POWER from
\code{\link{get_wth}}, but any data that has the properly named fields data
in them will work.
\itemize{
\item YYYYMMDD Date in YYYY-MM-DD format
\item DOY Numeric day of year, e.g. 1 - 365
\item T2M Mean daily temperature
\item T2MN Minimum daily temperature
\item T2MX Maximum daily temperature
\item RH2M Relative humidity
\item RAIN Precipitation
}}

\item{emergence}{Expected date of crop emergence}

\item{...}{Additional arguments, see \code{\link{SEIR}}}
}
\value{
An \pkg{epirice} \code{SEIR} object
}
\description{
A dynamic mechanistic simulation of leaf blast disease of rice, causal agent
\emph{Magnaporthe oryzae}.
The model is driven by daily weather data. Adapted from \pkg{cropsim} package
version 0.2.0-5 by Adam H. Sparks - USQ CCH.
Original model development: Serge Savary & Rene Pangga (IRRI).
Original \R implementation by Robert J. Hijmans, Rene Pangga, & Jorrel
Aunario (IRRI).
}
\examples{
\donttest{
# get weather for IRRI Zeigler Experiment Station in dry season 2000
wth <- get_wth(
  lonlat = c(121.25562, 14.6774),
  dates = c("2000-01-15", "2000-05-31")
)
lb <- predict_leaf_blast(wth, emergence = "2000-01-15")
plot(lb, type = 2)
}
}
\references{
Hwang B.K., Koh, Y.J., Chung, H.S., 1987. Effects of adult-plant
resistance on blast severity and yield of rice. Plant Disease 71:1035-1038.

Hemmi, T., Abe, T., Ikaya, J., and Inoue, Y. 1936. Studies on the
rice blast disease. IV. Relation of the environment to the development of
blast disease and physiologic specialization in the rice blast fungus.
Materials for Rural Improvement, Department of Agriculture and Forestry,
Japan No. 105, 145p.

Kato, H and Kozaka, T., 1974. Effect of temperature on lesion
enlargement and sporulation of \emph{Pyricularia oryzae} in rice leaves.
Phytopathology 64:828-830.

Torres, C.Q., 1986. Effect of plant age on the expression of
resistance to \emph{Pyricularia oryzae} Cav. in upland rice varieties. PhD
Thesis, University of the Philippines at Los Banos, 82 p.

El Refaei, M.I., 1977. Epidemiology of rice blast disease in the
tropics with special reference to the leaf wetness in relation to disease
development. PhD Thesis, Indian Agricultural Research Institute, New Delhi,
195 p.

Luo Wei-Hong, 1996. Simulation  and measurement of leaf wetness
formation in paddy rice crops. PhD Thesis, Wageningen Agricultural
University, 87 p.
}
\seealso{
\code{\link{predict_leaf_blast}},
\code{\link{predict_bacterial_blight}}, \code{\link{predict_brown_spot}},
\code{\link{predict_sheath_blight}}
}
\author{
Serge Savary, Ireneo Pangga, Robert Hijmans, Jorrel Khalil Aunario
}
