% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_leaf_rust.R
\name{predict_leaf_rust}
\alias{predict_leaf_rust}
\title{Predict wheat leaf or brown rust intensity}
\usage{
predict_leaf_rust(wth, emergence)
}
\arguments{
\item{wth}{Weather data with a daily time-step, normally \acronym{NASA}
\acronym{POWER} or \acronym{CHIRPS}/\acronym{CHIRTS} data from \code{\link[=get_wth]{get_wth()}},
but any\code{\link[base:data.frame]{base::data.frame()}} object that has the following properly named
columns in them will work.\tabular{rl}{
   \strong{Field Name} \tab \strong{Value} \cr
   \emph{YYYYMMDD} \tab Date as Year Month Day (ISO8601) \cr
   \emph{DOY} \tab Consecutive day of year, commonly called "Julian date" \cr
   \emph{TEMP} \tab Mean daily temperature (°C) \cr
   \emph{RAIN} \tab Mean daily rainfall (mm) \cr
   \emph{LAT} \tab \strong{Optional} latitude of weather observation. See LAT/LON Note. \cr
   \emph{LON} \tab \strong{Optional} longitude of weather observation. See LAT/LON Note. \cr
}}

\item{emergence}{Expected date of crop emergence}
}
\value{
A \code{\link[data.table:data.table]{data.table::data.table()}} of disease intensity and infection
sites. See \code{\link[=SEIR]{SEIR()}} for a full description of the column values.
}
\description{
A dynamic mechanistic simulation of leaf or brown rust disease of wheat,
causal agent \emph{Puccinia triticina}.  The model is driven by daily
weather data, which can easily be accessed using \code{\link[=get_wth]{get_wth()}} to download
weather data from \acronym{NASA} \acronym{POWER} using \CRANpkg{nasapower} or
\CRANpkg{chirps} for data from \acronym{CHIRPS} and \acronym{CHIRTS}.
}
\details{
The model represents site size as
4 \ifelse{html}{\out{mm<sup>2</sup>}}{\eqn{mm^2}} of a wheat plant's leaf.

Default values for this disease model are derived from Table 2 (Savary
\emph{et al.} 2015).
}
\note{
If the \code{wth} object provides \emph{LAT} and \emph{LON} columns, these will be included
in the output for mapping purposes. Both values must be present. These
columns are provided by default when using \code{\link[=get_wth]{get_wth()}}.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# get weather for IRRI Zeigler Experiment Station in wet season 2000
wth <- get_wth(
  lonlat = c(121.25562, 14.6774),
  dates = c("2000-06-30", "2000-12-31")
)
lr <- predict_leaf_rust(wth, emergence = "2000-07-01")
plot(x = lr$dates, y = lr$intensity, type = "l")
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=SEIR]{SEIR()}}

Other predict functions: 
\code{\link{predict_Septoria_tritici_blotch}()},
\code{\link{predict_bacterial_blight}()},
\code{\link{predict_brown_spot}()},
\code{\link{predict_leaf_blast}()},
\code{\link{predict_sheath_blight}()},
\code{\link{predict_tungro}()}
}
\author{
Adam H. Sparks, \email{adamhsparks@gmail.com}
}
\concept{predict functions}
