% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_sheath_blight.R
\name{predict_sheath_blight}
\alias{predict_sheath_blight}
\title{Predict rice sheath blight severity}
\usage{
predict_sheath_blight(wth, emergence, ...)
}
\arguments{
\item{wth}{Weather data with a daily time-step, normally NASA-POWER from
\code{\link{get_wth}}, but any \code{\link[base]{data.frame}} object that has
the following properly named columns in them will work.
\tabular{rl}{
\strong{YYYYMMDD}:\tab Date as Year Month Day (ISO8601).\cr
\strong{DOY}:\tab  Consecutive day of year, commonly called "Julian date".\cr
\strong{TM}:\tab Mean daily temperature (°C).\cr
\strong{TN}:\tab Minimum daily temperature (°C).\cr
\strong{TX}:\tab Maximum daily temperature (°C).\cr
\strong{TDEW}:\tab Mean daily dew point temperature (°C).\cr
\strong{RH}:\tab Mean daily relative humidity (\%).\cr
\strong{RAIN}:\tab Mean daily rainfall (mm).\cr
}}

\item{emergence}{Expected date of crop emergence}

\item{...}{Additional arguments, see \code{\link{SEIR}}}
}
\value{
A \code{\link[data.table]{data.table}} of disease severity and
infection sites. See \code{\link{SEIR}} for a full description of the
column values.
}
\description{
A dynamic mechanistic simulation of sheath blight disease of rice, causal
agent \emph{Rhizoctonia solani} AG1-1A Kühn.
The model is driven by daily weather data. Adapted from \pkg{cropsim} package
version 0.2.0-5 by Adam H. Sparks - USQ CCH.
Original model development: Serge Savary & Rene Pangga (IRRI).
Original \R implementation by Robert J. Hijmans, Rene Pangga, & Jorrel
Aunario (IRRI).
}
\examples{
\donttest{
# get weather for IRRI Zeigler Experiment Station in dry season 2000
wth <- get_wth(
  lonlat = c(121.25562, 14.6774),
  dates = c("2000-06-30", "2000-12-31")
)
sb <- predict_sheath_blight(wth, emergence = "2000-07-01")
plot(x = sb$dates, y = sb$severity, type = "l")
}
}
\references{
Savary, S., Castilla, N.P., Willocquet, L. 2001. Analysis of the
spatio-temporal structure of rice sheath blight epidemics in a farmer's
field. Plant Pathology 50:53-68.

Savary, S., Willocquet, L., Teng, P.S., 1997. Modelling sheath
blight epidemics on rice tillers. Agricultural Systems 55:359-384.

Castilla, N.P., Leano, R.M., Elazegui, F.A., Teng, P.S., Savary,
S., 1996. Effects of plant contacts, inoculation pattern, leaf wetness
regime, and nitrogen supply on inoculum efficiency in rice sheath blight.
Journal of Phytopathology 144:187-192.

Sharma, N.R., Teng, P.S., Olivares, F.M., 1990. Effect of rice
growth stage on sheath blight (ShB) development and yield loss. International
Rice Research Newsletter 15:19-20.

Hashiba, T. and Ijiri, T., 1989. Estimation of yield loss and
computerized forecasting system (BLIGHTAS) for rice sheath blight disease.
International Symposium on Tropical Agricultural Research: Crop losses due to
disease outbreaks in the tropics and countermeasures. Tropical Agricultural
Research Series (Japan) No. 22 pp. 163-171.

Tu, C.C., Chang, Y.C., Wang, C.W., 1979. Studies on the ecology
of \emph{Rhizoctonia solani}, the causal organism of rice sheath blight.
National Science Council Monthly, ROC 7:1208-1219.

Gross, M.K., Santini, J.B., Tikhonova, I. and Latin, R. 1998. The
influence of temperature and leaf wetness duration on infection of perennial
ryegrass by \emph{Rhizoctonia solani}. Plant Disease 82:1012-1016.
}
\seealso{
\itemize{
\item \code{\link{SEIR}},
\item \code{\link{predict_bacterial_blight}},
\item \code{\link{predict_brown_spot}},
\item \code{\link{predict_leaf_blast}},
\item \code{\link{predict_tungro}}
}
}
\author{
Serge Savary, Ireneo Pangga, Robert Hijmans, Jorrel Khalil Aunario
}
