% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tungro.R
\name{predict_tungro}
\alias{predict_tungro}
\title{Predict rice tungro area under the disease progress curve (AUDPC)}
\usage{
predict_tungro(wth, emergence, ...)
}
\arguments{
\item{wth}{Weather data with a daily time-step, normally NASA-POWER from
\code{\link{get_wth}}, but any data that has the properly named fields data
in them will work.
\itemize{
\item YYYYMMDD Date in YYYY-MM-DD format
\item DOY Numeric day of year, e.g. 1 - 365
\item T2M Mean daily temperature
\item T2MN Minimum daily temperature
\item T2MX Maximum daily temperature
\item RH2M Relative humidity
\item RAIN Precipitation
}}

\item{emergence}{Expected date of crop emergence}

\item{...}{Additional arguments, see \code{\link{SEIR}}}
}
\value{
An \pkg{epirice} \code{SEIR} object
}
\description{
A dynamic mechanistic simulation of tungro disease of rice, causal agents
\emph{Rice Tungro Spherical Virus} and \emph{Rice Tungro Baciliform Virus}.
The model is driven by daily weather data. Adapted from \pkg{cropsim} package
version 0.2.0-5 by Adam H. Sparks - USQ CCH.
Original model development: Serge Savary & Rene Pangga (IRRI).
Original \R implementation by Robert J. Hijmans, Rene Pangga, & Jorrel
Aunario (IRRI).
}
\examples{
\donttest{
# get weather for IRRI Zeigler Experiment Station in dry season 2000
wth <- get_wth(
  lonlat = c(121.25562, 14.6774),
  dates = c("2000-05-15", "2000-12-31")
)
tg <- predict_tungro(wth, emergence = "2000-05-18")
}
}
\references{
Tiongco, E.R., Cabunagan, R.C., Flores, Z.M., Hibino, H., and
Koganezawa, H., 1993. Serological monitoring of rice tungro disease
development in the field: its implication in disease management.
Plant Disease 77:877-882.

Rivera, C.T. and Ou, S.H., 1965. Leafhopper transmission of
tungro disease of rice. Plant Disease Reporter 49:127-131.

Ling, K.C., Palomar, M.K., 1966. Studies on rice plants infected
with the tungro virus at different ages. Philippines Agriculturist 50:165-177.

Ling, K.C., and Tiongco, E.R., Effect of temperature on the
transmission of rice tungro virus by \emph{Nephotettix virescens}.
Philippine Phytopathology 11:46-57.
}
\seealso{
\code{\link{predict_leaf_blast}},
\code{\link{predict_bacterial_blight}}, \code{\link{predict_brown_spot}},
\code{\link{predict_sheath_blight}}
}
\author{
Serge Savary, Ireneo Pangga, Robert Hijmans, Jorrel Khalil Aunario
}
