% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tungro.R
\name{predict_tungro}
\alias{predict_tungro}
\title{Predict tungro area under the disease progress curve (AUDPC)}
\usage{
predict_tungro(wth, emergence = "2000-05-15", ...)
}
\arguments{
\item{wth}{Weather data with a daily time-step, normally NASA-POWER, but any
data that has the following fields will work:
\describe{
  \item{YEAR}{Year in YYYY format}
  \item{DOY}{Numeric day of year, e.g. 1 - 365}
  \item{T2M}{Mean daily temperature}
  \item{T2MN}{Minimum daily temperature}
  \item{T2MX}{Maximum daily temperature}
  \item{RH2M}{Relative humidity}
  \item{RAIN}{Precipitation}
}}

\item{emergence}{Expected date of crop emergence}

\item{...}{Additional arguments, see \code{\link{SEIR}}}
}
\value{
A raster of AUDPC values for tungro based on given weather data
}
\description{
A dynamic mechanistic simulation of tungro disease of rice.
The model is driven by daily weather data. Adapted from cropsim package
version 0.2.0-5 by Adam H. Sparks - USQ CCH.
Original model development: Serge Savary & Rene Pangga (IRRI).
Original R implementation by Robert J. Hijmans, Rene Pangga, & Jorrel Aunario
(IRRI).
}
\examples{
wth <-
tg <- predict_tungro(wth, emergence = "2000-05-15")
plot(tg, type = 2)

}
\references{
Tiongco, E.R., Cabunagan, R.C., Flores, Z.M., Hibino, H., and
Koganezawa, H., 1993. Serological monitoring of rice tungro disease
development in the field: its implication in disease management.
Plant Disease 77:877-882.

Rivera, C.T. and Ou, S.H., 1965. Leafhopper transmission of
tungro disease of rice. Plant Disease Reporter 49:127-131.

Ling, K.C., Palomar, M.K., 1966. Studies on rice plants infected
with the tungro virus at different ages. Philippines Agriculturist 50:165-177.

Ling, K.C., and Tiongco, E.R., Effect of temperature on the
transmission of rice tungro virus by \emph{Nephotettix virescens}.
Philippine Phytopathology 11:46-57.
}
\seealso{
#' \code{\link{predict_leaf_blast}},
\code{\link{predict_bacterial_blight}}, \code{\link{predict_brown_spot}},
\code{\link{predict_sheath_blight}}
}
\author{
Serge Savary, Ireneo Pangga, Robert Hijmans, Jorrel Khalil Aunario
}
