% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR.R
\name{SEIR}
\alias{SEIR}
\title{Susceptible-Exposed-Infectious-Removed (SEIR) model}
\usage{
SEIR(
  wth,
  emergence,
  onset = 15,
  duration = 120,
  rhlim = 90,
  rainlim = 5,
  wetness = 0,
  init_sites,
  init_infection = 1,
  age_rc,
  tmp_rc,
  rh_rc,
  base_rc,
  latrans,
  inftrans,
  site_max,
  aggr,
  rr_physiol_senesc,
  rrg,
  senesc_type = 1
)
}
\arguments{
\item{wth}{a data frame of weather on a daily time-step containing data
with the following field names.
\itemize{
\item YYYYMMDD Date in YYYY-MM-DD format
\item DOY Numeric day of year, e.g. 1 - 365
\item T2M Mean daily temperature
\item T2MN Minimum daily temperature
\item T2MX Maximum daily temperature
\item RH2M Relative humidity
\item RAIN Precipitation
}}

\item{emergence}{expected date of plant emergence entered in \code{YYYY-MM-DD}
format (usually supplied through the disease model function)}

\item{onset}{expected number of days until the onset of disease after
emergence date}

\item{duration}{simulation duration}

\item{rhlim}{threshold to decide whether leaves are wet or not (usually
90 pct)}

\item{rainlim}{threshold to decide whether leaves are wet or not}

\item{wetness}{simulate RHmax or rain threshold (0) or leaf wetness duration
(1)}

\item{init_sites}{Number of initial infection sites}

\item{init_infection}{NA}

\item{age_rc}{NA}

\item{tmp_rc}{NA}

\item{rh_rc}{NA}

\item{base_rc}{corrected basic infection rate}

\item{latrans}{latent period}

\item{inftrans}{infectious period}

\item{site_max}{NA}

\item{aggr}{NA}

\item{rr_physiol_senesc}{relative rate of physiological senescence}

\item{rrg}{NA}

\item{senesc_type}{NA}
}
\value{
A \code{SEIR} \code{\link{class}} object that is a
\code{\link[data.table]{data.table}} that allows for simple plotting of the
contents
}
\description{
This function is used by specific disease models in EPIRICE to model disease
severity of several rice diseases.  It should be generic enough to port to
other pathosystems.
}
\details{
\code{SEIR} is called by the following specific disease models:
\code{\link{predict_leaf_blast}}, \code{\link{predict_bacterial_blight}},
\code{\link{predict_brown_spot}}, \code{\link{predict_sheath_blight}}
}
\examples{
\donttest{
# get weather for IRRI Zeigler Experiment Station in dry season 2000
wth <- get_wth(
  lonlat = c(121.25562, 14.6774),
  dates = c("2000-01-15", "2000-05-31")
)

# provide suitable values for brown spot severity
age_rc <-
  cbind(0:6 * 20, c(0.35, 0.35, 0.35, 0.47, 0.59, 0.71, 1.0))
tmp_rc <-
  cbind(15 + (0:5) * 5, c(0, 0.06, 1.0, 0.85, 0.16, 0))
rh_rc <- cbind(0:8 * 3,
             c(0, 0.12, 0.20, 0.38, 0.46, 0.60, 0.73, 0.87, 1.0))
emergence <- "2000-01-15"

x <- SEIR(
  wth = wth,
  emergence = emergence,
  age_rc = age_rc,
  tmp_rc = tmp_rc,
  rh_rc = rh_rc,
  base_rc = 0.61,
  latrans = 6,
  inftrans = 19,
  init_sites = 600,
  aggr = 1,
  site_max = 100000,
  rr_physiol_senesc = 0.01,
  rrg = 0.1
)
}
}
\author{
Serge Savary, Ireneo Pangga, Robert Hijmans, Jorrel Khalil Aunario,
Adam H. Sparks, Aji Sukarta
}
