% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR.R
\name{SEIR}
\alias{SEIR}
\title{Susceptible-Exposed-Infectious-Removed (SEIR) model framework}
\usage{
SEIR(
  wth,
  emergence,
  onset = 15,
  duration = 120,
  rhlim = 90,
  rainlim = 5,
  wetness_type = 0,
  H0,
  I0 = 1,
  RcA,
  RcT,
  RcW,
  RcOpt,
  l,
  i,
  Sx,
  a,
  RRS,
  RRG
)
}
\arguments{
\item{wth}{a data frame of weather on a daily time-step containing data with
the following field names.
\itemize{
\item YYYYMMDD Date in YYYY-MM-DD format
\item DOY Numeric day of year, e.g. 1 - 365
\item TM Mean daily temperature
\item TN Minimum daily temperature
\item TX Maximum daily temperature
\item TDEW Mean daily dewpoint temperature
\item RH Relative humidity
\item RAIN Precipitation
}}

\item{emergence}{expected date of plant emergence entered in \code{YYYY-MM-DD}
format}

\item{onset}{expected number of days until the onset of disease after
emergence date}

\item{duration}{simulation duration (growing season length)}

\item{rhlim}{threshold to decide whether leaves are wet or not (usually
90 \%)}

\item{rainlim}{threshold to decide whether leaves are wet or not}

\item{wetness_type}{simulate RHmax or rain threshold (0) or leaf wetness
duration (1)}

\item{H0}{Initial number of plant's healthy sites}

\item{I0}{Initial number of infective sites}

\item{RcA}{crop age curve for pathogen optimum}

\item{RcT}{temperature curve for pathogen optimum}

\item{RcW}{relative curve for pathogen optimum}

\item{RcOpt}{potential basic infection rate corrected for removals}

\item{l}{latent period}

\item{i}{infectious period}

\item{Sx}{maximum number of sites}

\item{a}{aggregation coefficient}

\item{RRS}{relative rate of physiological senescence}

\item{RRG}{relative rate of growth}
}
\value{
A \code{\link[data.table]{data.table}} containing the following
columns
\tabular{rl}{
\strong{simday}:\tab Zero indexed day of simulation that was run.\cr
\strong{dates}:\tab  Date of simulation.\cr
\strong{sites}:\tab Total number of sites present on day "x".\cr
\strong{latent}:\tab Number of latent sites present on day "x".\cr
\strong{infectious}:\tab Number of infectious sites present on day "x".\cr
\strong{removed}:\tab Number of removed sites present on day "x".\cr
\strong{senesced}:\tab Number of senesced sites present on day "x".\cr
\strong{rateinf}:\tab ...\cr
\strong{rtransfer}:\tab ...\cr
\strong{rgrowth}:\tab ...\cr
\strong{rsenesced}:\tab  ...\cr
\strong{diseased}:\tab  ...\cr
\strong{severity}:\tab  Disease severity or incidence (for tungro).\cr
}
}
\description{
This function is originally used by specific disease models in EPIRICE to
model disease severity of several rice diseases.  Given proper values it can
be used with other pathosystems as well.
}
\details{
\code{SEIR} is called by the following specific disease modelling
functions:
\itemize{
\item \code{\link{predict_bacterial_blight}},
\item \code{\link{predict_brown_spot}},
\item \code{\link{predict_leaf_blast}},
\item \code{\link{predict_sheath_blight}},
\item \code{\link{predict_tungro}}
}
}
\examples{
\donttest{
# get weather for IRRI Zeigler Experiment Station in dry season 2000
wth <- get_wth(
  lonlat = c(121.25562, 14.6774),
  dates = c("2000-06-30", "2000-12-31")
)

# provide suitable values for brown spot severity
RcA <-
  cbind(0:6 * 20, c(0.35, 0.35, 0.35, 0.47, 0.59, 0.71, 1.0))
RcT <-
  cbind(15 + (0:5) * 5, c(0, 0.06, 1.0, 0.85, 0.16, 0))
RcW <- cbind(0:8 * 3,
             c(0, 0.12, 0.20, 0.38, 0.46, 0.60, 0.73, 0.87, 1.0))
emergence <- "2000-07-15"

(SEIR(
  wth = wth,
  emergence = emergence,
  RcA = RcA,
  RcT = RcT,
  RcW = RcW,
  RcOpt = 0.61,
  l = 6,
  i = 19,
  H0 = 600,
  a = 1,
  Sx = 100000,
  RRS = 0.01,
  RRG = 0.1
))
}
}
\references{
Serge Savary, Andrew Nelson, Laetitia Willocquet, Ireneo Pangga
and Jorrel Aunario. Modeling and mapping potential epidemics of rice
diseases globally. Crop Protection, Volume 34, 2012, Pages 6-17, ISSN
0261-2194 DOI: \url{http://dx.doi.org/10.1016/j.cropro.2011.11.009}
}
\author{
Serge Savary, Ireneo Pangga, Robert Hijmans, Jorrel Khalil Aunario,
Adam H. Sparks, Aji Sukarta
}
