% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR.R
\name{SEIR}
\alias{SEIR}
\title{Susceptible-Exposed-Infectious-Removed (SEIR) Model}
\usage{
SEIR(wth, emergence, onset = 15, duration = 120, rhlim = 90,
rainlim = 5, wetness = 0, initSites, initInfection = 1, ageRc, tmpRc, rhRc,
baseRc, latrans, inftrans, siteMax, AGGR, RRPhysiolSenesc, RRG,
SenescType = 1)
}
\arguments{
\item{wth}{- daily weather data frame containing relative humidity (relh),
minimum temperature (tmin), and maximum temperature (tmax)}

\item{emergence}{- expected date of plant emergence}

\item{onset}{-expected number of days until the onset of disease after
emergence date}

\item{duration}{- simulation duration}

\item{rhlim}{-threshold to decide whether leaves are wet or not (usually
90\%)}

\item{rainlim}{- threshold to decide whether leaves are wet or not}

\item{wetness}{-simulate RHmax or rain threshold (0) or leaf wetness duration
(1)}

\item{init_sites}{-}

\item{init_infection}{-}

\item{age_rc}{-}

\item{tmp_rc}{-}

\item{rh_rc}{-}

\item{base_rc}{- corrected basic infection rate}

\item{latrans}{- latent period}

\item{inftrans}{- infectious period}

\item{site_max}{-}

\item{aggr}{-}

\item{rr_physiol_senesc}{- relative rate of physiological senescence}

\item{rrg}{-}

\item{senesc_type}{-}
}
\description{
This function is used by specific disease models in EPIRICE to model disease
severity.
}
\details{
SEIR is called by the following specific disease models:
\code{\link[epiRice]{leaf_blast}}, \code{\link[epiRice]{bacterial_blight}},
\code{\link[epiRice]{brown_spot}}, \code{\link[epiRice]{sheath_blight}},
\code{\link[epiRice]{tungro}}
}
\author{
Serge Savary, Ireneo Pangga, Robert Hijmans, Jorrel Khalil Aunario
}
