% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR.R
\name{SEIR}
\alias{SEIR}
\title{Susceptible-Exposed-Infectious-Removed (SEIR) model}
\usage{
SEIR(
  wth,
  emergence,
  onset = 15,
  duration = 120,
  rhlim = 90,
  rainlim = 5,
  wetness = 0,
  init_sites,
  init_infection = 1,
  age_rc,
  tmp_rc,
  rh_rc,
  base_rc,
  latrans,
  inftrans,
  site_max,
  aggr,
  rr_physiol_senesc,
  rrg,
  senesc_type = 1
)
}
\arguments{
\item{wth}{a data frame of weather on a daily time-step containing data
with the following field names.
\itemize{
\item YYYYMMDD Date in YYYY-MM-DD format
\item DOY Numeric day of year, e.g. 1 - 365
\item T2M Mean daily temperature
\item T2MN Minimum daily temperature
\item T2MX Maximum daily temperature
\item RH2M Relative humidity
\item RAIN Precipitation
}}

\item{emergence}{expected date of plant emergence entered in \code{YYYY-MM-DD}
format (usually supplied through the disease model function)}

\item{onset}{expected number of days until the onset of disease after
emergence date}

\item{duration}{simulation duration}

\item{rhlim}{threshold to decide whether leaves are wet or not (usually
90 pct)}

\item{rainlim}{threshold to decide whether leaves are wet or not}

\item{wetness}{simulate RHmax or rain threshold (0) or leaf wetness duration
(1)}

\item{init_sites}{NA}

\item{init_infection}{NA}

\item{age_rc}{NA}

\item{tmp_rc}{NA}

\item{rh_rc}{NA}

\item{base_rc}{corrected basic infection rate}

\item{latrans}{latent period}

\item{inftrans}{infectious period}

\item{site_max}{NA}

\item{aggr}{NA}

\item{rr_physiol_senesc}{relative rate of physiological senescence}

\item{rrg}{NA}

\item{senesc_type}{NA}
}
\description{
This function is used by specific disease models in EPIRICE to model disease
severity.
}
\details{
SEIR is called by the following specific disease models:
\code{\link{predict_leaf_blast}}, \code{\link{predict_bacterial_blight}},
\code{\link{predict_brown_spot}}, \code{\link{predict_sheath_blight}}
}
\author{
Serge Savary, Ireneo Pangga, Robert Hijmans, Jorrel Khalil Aunario,
Adam H Sparks, Aji Sukarta
}
